Added the queen + made logic.rs it's own seperate module
So making the queen was extremely easy, I literally just copied and pasted the code for the bishop and rook, and added some conditionals to make sure the queen was moving like a bishop or a rook.

As for my reasoning on making logic.rs it's own entirely seperate module, I read somewhere that making modules that work entirely independently hould help increase Rust compile times, plus it's much easier to read and understand that the drawing code is in main.rs, and the code for the game's logic is in logic.rs.