Added Knight movement
The knight is honestly the simplest piece to program so far. I just manually wrote some code for generating the 8 possible moves a knight can make, and just check to see that the player is making one of them.

I also just finished up with adding all possible enums to the list, since the only real piece left is the King.